# open-api-doc/api-documentation.yml
{"openapi":"3.0.1","info":{"title":"Weather Service API Documentation","description":"API endpoints to query weather using OpenWeatherMap API.","contact":{"name":"Neel Priyadarshi","email":"neel.priyadarshi18@gmail.com"},"version":"1.0"},"servers":[{"url":"http://bytemetwice.ninja","description":"Development"}],"tags":[{"name":"City Search Analytics","description":"APIs to Get Analytics On City Search Data."},{"name":"Home","description":"Redirect to Swagger UI"},{"name":"Actuator","description":"Monitor and interact","externalDocs":{"description":"Spring Boot Actuator Web API Documentation","url":"https://docs.spring.io/spring-boot/docs/current/actuator-api/html/"}},{"name":"Current Weather APIs","description":"APIs to retrieve the current weather of the city based on provided parameters."}],"paths":{"/api/v1/current/weather/zip":{"get":{"tags":["Current Weather APIs"],"summary":"Get current weather by zip","description":"Get current weather by zip code; Country code is optional","operationId":"getCurrentWeatherByZip","parameters":[{"name":"zipCode","in":"query","description":"Zip; Required Field","required":true,"schema":{"type":"string"}},{"name":"countryCode","in":"query","description":"Country Code; Optional Field","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WeatherResponse"}}}},"404":{"description":"Weather information not found for specified parameters"},"400":{"description":"Bad Request, Please verify the request parameters and try again"}}},"post":{"tags":["Current Weather APIs"],"summary":"POST as GET: Get current weather by zip request model","description":"POST as GET: Get current weather by zip code and country code; Country code is optional","operationId":"getCurrentWeatherByZipRequestModel","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ZipCodeRequestModel"}}},"required":true},"responses":{"404":{"description":"Weather information not available for specified parameters"},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WeatherResponse"}}}},"400":{"description":"Bad Request, Please verify the request parameters and try again"}}}},"/api/v1/current/weather/coordinates":{"get":{"tags":["Current Weather APIs"],"summary":"Get current weather by coordinates","description":"Get current weather by latitude and longitude; Required Fields","operationId":"getCurrentWeatherByCoordinates","parameters":[{"name":"latitude","in":"query","description":"Latitude; Required Field","required":true,"schema":{"type":"number","format":"double"}},{"name":"longitude","in":"query","description":"Longitude; Required Field","required":true,"schema":{"type":"number","format":"double"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WeatherResponse"}}}},"404":{"description":"Weather information not found for specified parameters"},"400":{"description":"Bad Request, Please verify the request parameters and try again"}}},"post":{"tags":["Current Weather APIs"],"summary":"POST as GET: Get current weather by coordinates request model","description":"POST as GET: Get current weather by coordinates; Latitude and Longitude are required fields","operationId":"getCurrentWeatherByCoordinateRequestModel","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CoordinateRequestModel"}}},"required":true},"responses":{"404":{"description":"Weather information not available for specified parameters"},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WeatherResponse"}}}},"400":{"description":"Bad Request, Please verify the request parameters and try again"}}}},"/api/v1/current/weather/city":{"get":{"tags":["Current Weather APIs"],"summary":"Get current weather by city","description":"Get current weather by city ID or city name; State code and country code are optional","operationId":"getCurrentWeatherByCity","parameters":[{"name":"id","in":"query","description":"ID of the City; Required Field","required":false,"schema":{"type":"string"}},{"name":"name","in":"query","description":"Name of City; Required Field","required":false,"schema":{"type":"string"}},{"name":"stateCode","in":"query","description":"State Code; Optional Field","required":false,"schema":{"type":"string"}},{"name":"countryCode","in":"query","description":"Country Code; Optional Field","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WeatherResponse"}}}},"404":{"description":"Weather information not found for specified parameters"},"400":{"description":"Bad Request, Please verify the request parameters and try again"}}},"post":{"tags":["Current Weather APIs"],"summary":"POST as GET: Get current weather by city request model","description":"POST as GET: Get current weather by city name, state code and country code; State code and country code are optional","operationId":"getCurrentWeatherByCityRequestModel","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CityRequestModel"}}},"required":true},"responses":{"404":{"description":"Weather information not available for specified parameters"},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WeatherResponse"}}}},"400":{"description":"Bad Request, Please verify the request parameters and try again"}}}},"/api/v1/tracker/city":{"get":{"tags":["City Search Analytics"],"summary":"Get City Search Tracker; If City Name Is Not Provided All Cities Tracker Will Be Returned","description":"Get City Search Tracker; If City Name Is Not Provided All Cities Tracker Will Be Returned","operationId":"getCityTracker","parameters":[{"name":"name","in":"query","description":"City Name; Optional Field","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","description":"Page Number; Optional Field","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"pageSize","in":"query","description":"Page Size: No of Entries per Page; Optional Field","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CityTrackerResponse"}}}}}}},"/actuator":{"get":{"tags":["Actuator"],"summary":"Actuator root web endpoint","operationId":"links","responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object","additionalProperties":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/Link"}}}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object","additionalProperties":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/Link"}}}},"application/json":{"schema":{"type":"object","additionalProperties":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/Link"}}}}}}}}},"/actuator/feature":{"get":{"tags":["Actuator"],"summary":"Actuator web endpoint 'feature'","operationId":"feature","responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}},"application/json":{"schema":{"type":"object"}}}}}}},"/actuator/feature/{featureName}":{"get":{"tags":["Actuator"],"summary":"Actuator web endpoint 'feature-featureName'","operationId":"feature-featureName","parameters":[{"name":"featureName","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}},"application/json":{"schema":{"type":"object"}}}}}}},"/":{"get":{"tags":["Home"],"operationId":"home","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RedirectView"}}}}}}}},"components":{"schemas":{"ZipCodeRequestModel":{"required":["zipCode"],"type":"object","properties":{"zipCode":{"pattern":"^[a-zA-Z0-9]{1,10}$","type":"string"},"countryCode":{"pattern":"^[a-zA-Z]{2}$","type":"string"}}},"Clouds":{"type":"object","properties":{"all":{"type":"integer","format":"int32"}}},"Coord":{"type":"object","properties":{"lon":{"type":"number","format":"double"},"lat":{"type":"number","format":"double"}}},"Main":{"type":"object","properties":{"temp":{"type":"number","format":"double"},"feels_like":{"type":"number","format":"double"},"temp_min":{"type":"number","format":"double"},"temp_max":{"type":"number","format":"double"},"pressure":{"type":"integer","format":"int32"},"humidity":{"type":"integer","format":"int32"},"sea_level":{"type":"integer","format":"int32"},"grnd_level":{"type":"integer","format":"int32"}}},"Rain":{"type":"object","properties":{"1h":{"type":"number","format":"double"}}},"Sys":{"type":"object","properties":{"type":{"type":"integer","format":"int32"},"id":{"type":"integer","format":"int32"},"country":{"type":"string"},"sunrise":{"type":"integer","format":"int32"},"sunset":{"type":"integer","format":"int32"}}},"Weather":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"main":{"type":"string"},"description":{"type":"string"},"icon":{"type":"string"}}},"WeatherModel":{"type":"object","properties":{"coord":{"$ref":"#/components/schemas/Coord"},"weather":{"type":"array","items":{"$ref":"#/components/schemas/Weather"}},"base":{"type":"string"},"main":{"$ref":"#/components/schemas/Main"},"visibility":{"type":"integer","format":"int32"},"wind":{"$ref":"#/components/schemas/Wind"},"rain":{"$ref":"#/components/schemas/Rain"},"clouds":{"$ref":"#/components/schemas/Clouds"},"dt":{"type":"integer","format":"int64"},"sys":{"$ref":"#/components/schemas/Sys"},"timezone":{"type":"integer","format":"int64"},"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"cod":{"type":"integer","format":"int32"}}},"WeatherResponse":{"type":"object","properties":{"currentWeather":{"$ref":"#/components/schemas/WeatherModel"},"message":{"type":"string"}}},"Wind":{"type":"object","properties":{"speed":{"type":"number","format":"double"},"deg":{"type":"integer","format":"int32"},"gust":{"type":"number","format":"double"}}},"CoordinateRequestModel":{"required":["latitude","longitude"],"type":"object","properties":{"latitude":{"maximum":90,"exclusiveMaximum":false,"minimum":-90,"exclusiveMinimum":false,"type":"number","format":"double"},"longitude":{"maximum":180,"exclusiveMaximum":false,"minimum":-180,"exclusiveMinimum":false,"type":"number","format":"double"}}},"CityRequestModel":{"type":"object","properties":{"id":{"pattern":"^\\d+$","type":"string"},"name":{"pattern":"^[a-zA-Z]+$","type":"string"},"stateCode":{"pattern":"^[a-zA-Z]{2}$","type":"string"},"countryCode":{"pattern":"^[a-zA-Z]{2}$","type":"string"}}},"CityTrackerResponse":{"type":"object","properties":{"name":{"type":"string"},"idCount":{"type":"integer","format":"int64"},"coordinateCount":{"type":"integer","format":"int64"},"nameCount":{"type":"integer","format":"int64"},"zipCodeCount":{"type":"integer","format":"int64"},"totalCounter":{"type":"integer","format":"int64"},"message":{"type":"string"}}},"Link":{"type":"object","properties":{"href":{"type":"string"},"templated":{"type":"boolean"}}},"ApplicationContext":{"type":"object","properties":{"parent":{"$ref":"#/components/schemas/ApplicationContext"},"id":{"type":"string"},"displayName":{"type":"string"},"applicationName":{"type":"string"},"startupDate":{"type":"integer","format":"int64"},"autowireCapableBeanFactory":{"$ref":"#/components/schemas/AutowireCapableBeanFactory"},"environment":{"$ref":"#/components/schemas/Environment"},"beanDefinitionCount":{"type":"integer","format":"int32"},"beanDefinitionNames":{"type":"array","items":{"type":"string"}},"parentBeanFactory":{"$ref":"#/components/schemas/BeanFactory"},"classLoader":{"type":"object","properties":{"name":{"type":"string"},"registeredAsParallelCapable":{"type":"boolean"},"parent":{"type":"object","properties":{"name":{"type":"string"},"registeredAsParallelCapable":{"type":"boolean"},"unnamedModule":{"type":"object","properties":{"name":{"type":"string"},"classLoader":{"type":"object","properties":{"name":{"type":"string"},"registeredAsParallelCapable":{"type":"boolean"},"definedPackages":{"type":"array","items":{"type":"object","properties":{"name":{"type":"string"},"annotations":{"type":"array","items":{"type":"object"}},"declaredAnnotations":{"type":"array","items":{"type":"object"}},"sealed":{"type":"boolean"},"specificationTitle":{"type":"string"},"specificationVersion":{"type":"string"},"specificationVendor":{"type":"string"},"implementationTitle":{"type":"string"},"implementationVersion":{"type":"string"},"implementationVendor":{"type":"string"}}}},"defaultAssertionStatus":{"type":"boolean","writeOnly":true}}},"descriptor":{"type":"object","properties":{"open":{"type":"boolean"},"automatic":{"type":"boolean"}}},"named":{"type":"boolean"},"annotations":{"type":"array","items":{"type":"object"}},"declaredAnnotations":{"type":"array","items":{"type":"object"}},"packages":{"uniqueItems":true,"type":"array","items":{"type":"string"}},"layer":{"type":"object"}}},"definedPackages":{"type":"array","items":{"type":"object","properties":{"name":{"type":"string"},"annotations":{"type":"array","items":{"type":"object"}},"declaredAnnotations":{"type":"array","items":{"type":"object"}},"sealed":{"type":"boolean"},"specificationTitle":{"type":"string"},"specificationVersion":{"type":"string"},"specificationVendor":{"type":"string"},"implementationTitle":{"type":"string"},"implementationVersion":{"type":"string"},"implementationVendor":{"type":"string"}}}},"defaultAssertionStatus":{"type":"boolean","writeOnly":true}}},"unnamedModule":{"type":"object","properties":{"name":{"type":"string"},"classLoader":{"type":"object","properties":{"name":{"type":"string"},"registeredAsParallelCapable":{"type":"boolean"},"definedPackages":{"type":"array","items":{"type":"object","properties":{"name":{"type":"string"},"annotations":{"type":"array","items":{"type":"object"}},"declaredAnnotations":{"type":"array","items":{"type":"object"}},"sealed":{"type":"boolean"},"specificationTitle":{"type":"string"},"specificationVersion":{"type":"string"},"specificationVendor":{"type":"string"},"implementationTitle":{"type":"string"},"implementationVersion":{"type":"string"},"implementationVendor":{"type":"string"}}}},"defaultAssertionStatus":{"type":"boolean","writeOnly":true}}},"descriptor":{"type":"object","properties":{"open":{"type":"boolean"},"automatic":{"type":"boolean"}}},"named":{"type":"boolean"},"annotations":{"type":"array","items":{"type":"object"}},"declaredAnnotations":{"type":"array","items":{"type":"object"}},"packages":{"uniqueItems":true,"type":"array","items":{"type":"string"}},"layer":{"type":"object"}}},"definedPackages":{"type":"array","items":{"type":"object","properties":{"name":{"type":"string"},"annotations":{"type":"array","items":{"type":"object"}},"declaredAnnotations":{"type":"array","items":{"type":"object"}},"sealed":{"type":"boolean"},"specificationTitle":{"type":"string"},"specificationVersion":{"type":"string"},"specificationVendor":{"type":"string"},"implementationTitle":{"type":"string"},"implementationVersion":{"type":"string"},"implementationVendor":{"type":"string"}}}},"defaultAssertionStatus":{"type":"boolean","writeOnly":true}}}}},"AutowireCapableBeanFactory":{"type":"object"},"BeanFactory":{"type":"object"},"Environment":{"type":"object","properties":{"activeProfiles":{"type":"array","items":{"type":"string"}},"defaultProfiles":{"type":"array","items":{"type":"string"}}}},"FilterRegistration":{"type":"object","properties":{"servletNameMappings":{"type":"array","items":{"type":"string"}},"urlPatternMappings":{"type":"array","items":{"type":"string"}},"name":{"type":"string"},"className":{"type":"string"},"initParameters":{"type":"object","additionalProperties":{"type":"string"}}}},"HttpStatusCode":{"type":"object","properties":{"error":{"type":"boolean"},"is4xxClientError":{"type":"boolean"},"is5xxServerError":{"type":"boolean"},"is1xxInformational":{"type":"boolean"},"is2xxSuccessful":{"type":"boolean"},"is3xxRedirection":{"type":"boolean"}}},"JspConfigDescriptor":{"type":"object","properties":{"taglibs":{"type":"array","items":{"$ref":"#/components/schemas/TaglibDescriptor"}},"jspPropertyGroups":{"type":"array","items":{"$ref":"#/components/schemas/JspPropertyGroupDescriptor"}}}},"JspPropertyGroupDescriptor":{"type":"object","properties":{"buffer":{"type":"string"},"elIgnored":{"type":"string"},"errorOnELNotFound":{"type":"string"},"pageEncoding":{"type":"string"},"scriptingInvalid":{"type":"string"},"isXml":{"type":"string"},"includePreludes":{"type":"array","items":{"type":"string"}},"includeCodas":{"type":"array","items":{"type":"string"}},"deferredSyntaxAllowedAsLiteral":{"type":"string"},"trimDirectiveWhitespaces":{"type":"string"},"errorOnUndeclaredNamespace":{"type":"string"},"defaultContentType":{"type":"string"},"urlPatterns":{"type":"array","items":{"type":"string"}}}},"RedirectView":{"type":"object","properties":{"applicationContext":{"$ref":"#/components/schemas/ApplicationContext"},"servletContext":{"$ref":"#/components/schemas/ServletContext"},"contentType":{"type":"string"},"requestContextAttribute":{"type":"string"},"staticAttributes":{"type":"object","additionalProperties":{"type":"object"}},"exposePathVariables":{"type":"boolean"},"exposeContextBeansAsAttributes":{"type":"boolean","writeOnly":true},"exposedContextBeanNames":{"type":"array","writeOnly":true,"items":{"type":"string"}},"beanName":{"type":"string"},"url":{"type":"string"},"contextRelative":{"type":"boolean","writeOnly":true},"http10Compatible":{"type":"boolean","writeOnly":true},"exposeModelAttributes":{"type":"boolean","writeOnly":true},"encodingScheme":{"type":"string","writeOnly":true},"statusCode":{"$ref":"#/components/schemas/HttpStatusCode"},"expandUriTemplateVariables":{"type":"boolean","writeOnly":true},"propagateQueryParams":{"type":"boolean","writeOnly":true},"hosts":{"type":"array","items":{"type":"string"}},"redirectView":{"type":"boolean"},"propagateQueryProperties":{"type":"boolean"},"attributes":{"type":"object","additionalProperties":{"type":"string"},"writeOnly":true},"attributesCSV":{"type":"string","writeOnly":true},"attributesMap":{"type":"object","additionalProperties":{"type":"object"}}}},"ServletContext":{"type":"object","properties":{"classLoader":{"type":"object","properties":{"name":{"type":"string"},"registeredAsParallelCapable":{"type":"boolean"},"definedPackages":{"type":"array","items":{"type":"object","properties":{"name":{"type":"string"},"annotations":{"type":"array","items":{"type":"object"}},"declaredAnnotations":{"type":"array","items":{"type":"object"}},"sealed":{"type":"boolean"},"specificationTitle":{"type":"string"},"specificationVersion":{"type":"string"},"specificationVendor":{"type":"string"},"implementationTitle":{"type":"string"},"implementationVersion":{"type":"string"},"implementationVendor":{"type":"string"}}}},"defaultAssertionStatus":{"type":"boolean","writeOnly":true}}},"majorVersion":{"type":"integer","format":"int32"},"minorVersion":{"type":"integer","format":"int32"},"sessionCookieConfig":{"$ref":"#/components/schemas/SessionCookieConfig"},"virtualServerName":{"type":"string"},"attributeNames":{"type":"object"},"contextPath":{"type":"string"},"initParameterNames":{"type":"object"},"sessionTimeout":{"type":"integer","format":"int32"},"effectiveMajorVersion":{"type":"integer","format":"int32"},"effectiveMinorVersion":{"type":"integer","format":"int32"},"serverInfo":{"type":"string"},"servletContextName":{"type":"string"},"servletRegistrations":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/ServletRegistration"}},"filterRegistrations":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/FilterRegistration"}},"sessionTrackingModes":{"uniqueItems":true,"type":"array","writeOnly":true,"items":{"type":"string","enum":["COOKIE","URL","SSL"]}},"defaultSessionTrackingModes":{"uniqueItems":true,"type":"array","items":{"type":"string","enum":["COOKIE","URL","SSL"]}},"effectiveSessionTrackingModes":{"uniqueItems":true,"type":"array","items":{"type":"string","enum":["COOKIE","URL","SSL"]}},"jspConfigDescriptor":{"$ref":"#/components/schemas/JspConfigDescriptor"},"requestCharacterEncoding":{"type":"string"},"responseCharacterEncoding":{"type":"string"}}},"ServletRegistration":{"type":"object","properties":{"mappings":{"type":"array","items":{"type":"string"}},"runAsRole":{"type":"string"},"name":{"type":"string"},"className":{"type":"string"},"initParameters":{"type":"object","additionalProperties":{"type":"string"}}}},"SessionCookieConfig":{"type":"object","properties":{"name":{"type":"string"},"path":{"type":"string"},"attributes":{"type":"object","additionalProperties":{"type":"string"}},"comment":{"type":"string","deprecated":true},"maxAge":{"type":"integer","format":"int32"},"domain":{"type":"string"},"httpOnly":{"type":"boolean"},"secure":{"type":"boolean"}}},"TaglibDescriptor":{"type":"object","properties":{"taglibURI":{"type":"string"},"taglibLocation":{"type":"string"}}}}}}